# small performance improvement by using unsafe code