--- a/src/Solnet.Programs/Solnet.Programs.csproj
+++ b/src/Solnet.Programs/Solnet.Programs.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
   <ItemGroup>