> ⚠️ NOTE: Use notes like this to emphasize something important about the PR.
>  
>  This could include other PRs this PR is built on top of; API breaking changes; reasons for why the PR is on hold; or anything else you would like to draw attention to.

| Status  | Type  | ⚠️ Core Change | Issue |
| :---: | :---: | :---: | :--: |
| Ready | Optimization | No | NO|

## Problem

A  tiny performance gain in serialization utility ( see [this](https://stackoverflow.com/questions/4980881/what-is-fastest-way-to-convert-bool-to-byte)).


## Solution

using `unsafe` code without explicit converting and also removing branching.


## Before & After Performance

```
BenchmarkDotNet=v0.13.1, OS=pop 21.10
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK=6.0.102
  [Host]     : .NET 6.0.2 (6.0.222.6406), X64 RyuJIT
  DefaultJob : .NET 6.0.2 (6.0.222.6406), X64 RyuJIT


|                    Method |     Mean |   Error |  StdDev | Allocated |
|-------------------------- |---------:|--------:|--------:|----------:|
| explicit_branching_assign | 504.0 ns | 2.01 ns | 1.78 ns |         - |
|             unsafe_assing | 205.7 ns | 0.99 ns | 0.92 ns |         - |
```
## Benchmark Code

```csharp


using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

namespace Benchmark
{
    [MemoryDiagnoser]
    public  class Benchmark
    {
        private static byte _dummy;
        [Benchmark]
        public void explicit_branching_assign()
        {
            byte result = 0;
            var _the_bool = DateTimeOffset.Now.ToUnixTimeMilliseconds() % 2 ==0;
            for (var i = 0; i < 500; i++)
            {
                result = _the_bool ? (byte) (1) : (byte) 0;
            }
            _dummy = result ;
        }

        [Benchmark]
        public  unsafe void unsafe_assing(){
            byte result = 0;
            var _the_bool = DateTimeOffset.Now.ToUnixTimeMilliseconds() % 2 ==0;
            for (var i = 0; i < 500; i++)
            {
                result = *((byte*)(&_the_bool));
            }
            _dummy = result ;
        }
    }
    public static class MainClass
    {
        public static void Main()
        {
            var result = BenchmarkRunner.Run<Benchmark>();
        }
    }
}

```