@tiago18c 2022-03-01
The change looks good.
Just one minor thing: please revert the line ending changes.

ps: also, bump the version to 6.0.4 and I'll merge, the github actions will do the rest :)