@Neo-vortex 2022-03-01
I am a bit confused here!
Files in Utilities namespace are LF ending, while some others like SystemPorgram.cs is CRLF
Now I changed it back to CRLF i think.
is it ok or I change it to LF ?