@tiago18c 2022-03-01
The idea was to revert just this file in order to keep the PR simple. 
We wanted to do a global reformat in our v1, but we forgot about it.

Regarding the line endings, is there any way to force a specific line ending per project? Been looking into it, but so far I can't find anything. We're using windows, linux and macos and it would be great to have a way to force line endings to be equal across