@Neo-vortex 2022-03-01
probably the best way would be to let everyone works with their own native line ending.but configure `git` to do `autocrlf`.
You can also configure `.gitattributes` like this:
```
*.cs text eol=crlf
```
to force line ending. but  the IDE on Mac and Linux should support CRLF.
Read more [here](https://docs.github.com/en/get-started/getting-started-with-git/configuring-git-to-handle-line-endings)