@tiago18c 2022-03-01
Just found out something:

`dotnet format` doesn't respect `.editorconfig` parameter `end_of_line` if run under Windows, but does using MacOS.

I'll open a new PR later to run `dotnet format` and force `crlf` on all files :)